+++ main/confloader/external/external.go
@@ -6,22 +6,20 @@ import (
 	"bytes"
 	"io"
 	"io/ioutil"
-	"net/http"
-	"net/url"
 	"os"
 	"strings"
-	"time"
 
-	"v2ray.com/core/common/buf"
 	"v2ray.com/core/common/platform/ctlcmd"
+	"v2ray.com/core/infra/conf"
 	"v2ray.com/core/main/confloader"
 )
 
+// ConfigLoader is the protobuf config loader
 func ConfigLoader(arg string) (out io.Reader, err error) {
 	var data []byte
 	switch {
 	case strings.HasPrefix(arg, "http://"), strings.HasPrefix(arg, "https://"):
-		data, err = FetchHTTPContent(arg)
+		data, err = conf.FetchHTTPContent(arg)
 
 	case arg == "stdin:":
 		data, err = ioutil.ReadAll(os.Stdin)
@@ -35,41 +33,6 @@ func ConfigLoader(arg string) (out io.Reader, err error) {
 	}
 	out = bytes.NewBuffer(data)
 	return
-}
-
-func FetchHTTPContent(target string) ([]byte, error) {
-	parsedTarget, err := url.Parse(target)
-	if err != nil {
-		return nil, newError("invalid URL: ", target).Base(err)
-	}
-
-	if s := strings.ToLower(parsedTarget.Scheme); s != "http" && s != "https" {
-		return nil, newError("invalid scheme: ", parsedTarget.Scheme)
-	}
-
-	client := &http.Client{
-		Timeout: 30 * time.Second,
-	}
-	resp, err := client.Do(&http.Request{
-		Method: "GET",
-		URL:    parsedTarget,
-		Close:  true,
-	})
-	if err != nil {
-		return nil, newError("failed to dial to ", target).Base(err)
-	}
-	defer resp.Body.Close()
-
-	if resp.StatusCode != 200 {
-		return nil, newError("unexpected HTTP status code: ", resp.StatusCode)
-	}
-
-	content, err := buf.ReadAllToBytes(resp.Body)
-	if err != nil {
-		return nil, newError("failed to read HTTP response").Base(err)
-	}
-
-	return content, nil
 }
 
 // ExtConfigLoader calls v2ctl to load config
